trying to fix findLobby() maaaaaaaaaan

+++ Implementation/django_template/social_app/lobby.py
@@ -5,7 +5,6 @@ from .models import Player, Match, WaitingList
 from .kerstin_utils import *
 
 
-
 def addHostLobby(request):
     if not request.user.is_authenticated:
         return HttpResponse(f'user not signed in')
@@ -13,12 +12,16 @@ def addHostLobby(request):
         return HttpResponse(f'user is not a player')
 
     player: Player = request.user.player
-    try:
-        WaitingList.objects.get(player)
-        return HttpResponse(f'user already in HostLobby')
-    except ObjectDoesNotExist:
-        WaitingList(player).save()
-        return HttpResponse(f'0:')
+
+    count = player.waitinghost.all().count()
+    if count >= 1:
+        response = f'1: Player already in list'
+        return HttpResponse(response)
+    else:
+        WaitingList(waitinghost=player).save()
+        response = f'0:'
+        return HttpResponse(response)
+
 
 def findLobby(request):
     if request.method != 'POST':
@@ -27,13 +30,17 @@ def findLobby(request):
         return HttpResponse(not_a_player_message)
 
     player: Player = request.user.player
-    try:
-        host : Player = WaitingList.objects.first()
-        Match(host= host, joined_player= player).save()
+    waiters = WaitingList.objects.all()
+    print(waiters)
+    if waiters.count() >= 1:
+        host: Player = waiters.first()
+        print(host)
+        Match(host=host, joined_player=player).save()
         host.delete()
         return HttpResponse(f'0:' + f'{host.user.username}')
-    except ObjectDoesNotExist:
+    else:
         return HttpResponse(f'1:')
+
 
 def wait(request):
     if request.method != 'POST':
@@ -41,10 +48,10 @@ def wait(request):
     if not hasattr(request.user, 'player'):
         return HttpResponse(not_a_player_message)
 
-    player : Player = request.user.player
+    player: Player = request.user.player
 
     try:
-        guest : Player = Match.objects.get(host=player).joined_player
+        guest: Player = Match.objects.get(host=player).joined_player
         return HttpResponse(f'0:' + f'{guest.user.username}')
     except:
         return HttpResponse(f"1: waitin")
@@ -58,7 +65,7 @@ def setJoinedReady(request):
 
     player: Player = request.user.player
 
-    match : Match = Match.objects.get(joined_player=player)
+    match: Match = Match.objects.get(joined_player=player)
     match.guest_ready = True
     match.save()
     return HttpResponse(f'0:')
@@ -79,15 +86,16 @@ def startGame(request):
     else:
         return HttpResponse(f'lobby not ready')
 
+
 def leaveLobby(request):
     if request.method != 'POST':
         return HttpResponse(wrong_method_message)
     if not hasattr(request.user, 'player'):
         return HttpResponse(not_a_player_message)
 
-    player : Player = request.user.player
+    player: Player = request.user.player
 
-    if(isHost(player)):
+    if (isHost(player)):
         try:
             wait = WaitingList.objects.get(waitinghost=player)
             wait.delete()
@@ -102,12 +110,14 @@ def leaveLobby(request):
         Match.objects.filter(joined_player=player).delete()
         return HttpResponse(f"1: Joined Player left lobby -> host in waitinglist")
 
-def isHost(player : Player):
+
+def isHost(player: Player):
     try:
         Match.objects.get(host=player)
         return True
     except ObjectDoesNotExist:
         return False
+
 
 def isHost(request):
     if request.method != 'POST':
@@ -115,14 +125,15 @@ def isHost(request):
     if not hasattr(request.user, 'player'):
         return HttpResponse(not_a_player_message)
 
-    player : Player = request.user.player
-    try:
+    player: Player = request.user.player
+    count = player.host.all().count()
+    if count >= 1:
         WaitingList.objects.get(player)
         return HttpResponse(f'0:')
-    except ObjectDoesNotExist:
+    else:
         return HttpResponse(f'1:')
 
-#def checkIfFriend(request):
+# def checkIfFriend(request):
 #    if not request.user.is_authenticated:
 #        return HttpResponse(f'user not signed in')
 #    if request.method != 'GET':
